pyproject.toml
[project]
name = "vision-1m"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "openai>=2.29.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.44.1",
    "tavily-python>=0.7.23",
]
